extracted the upload API
